Split out item filter types
